Redesign GUI to output wiki revision results
This commit refactored the GUI code to be cleaner and added functionality
to the search button. Now our project outputs the latest revisions of a
specified Wiki page in reverse chronological order when the user inputs
the name of a Wiki page and presses search. The UserInterface class was
deleted because the redesign made it unnecessary.